Rearrange repos.
