--- src/parsers/character_parsers.rs
@@ -1,38 +1,0 @@
-use parcel::{MatchStatus, Parser};
-
-pub fn any<'a>() -> impl Parser<'a, &'a str, char> {
-    move |input: &'a str| match input.chars().next() {
-        Some(next) => Ok(MatchStatus::Match((&input[1..], next))),
-        _ => Ok(MatchStatus::NoMatch(input)),
-    }
-}
-
-pub fn alphabetic<'a>() -> impl Parser<'a, &'a str, char> {
-    move |input: &'a str| match input.chars().next() {
-        Some(next) if next.is_alphabetic() => Ok(MatchStatus::Match((&input[1..], next))),
-        _ => Ok(MatchStatus::NoMatch(input)),
-    }
-}
-
-pub fn alphabetic_or_dash<'a>() -> impl Parser<'a, &'a str, char> {
-    move |input: &'a str| match input.chars().next() {
-        Some(next) if next.is_alphabetic() || next == '-' || next == '_' => {
-            Ok(MatchStatus::Match((&input[1..], next)))
-        }
-        _ => Ok(MatchStatus::NoMatch(input)),
-    }
-}
-
-pub fn character<'a>(expected: char) -> impl Parser<'a, &'a str, char> {
-    move |input: &'a str| match input.chars().next() {
-        Some(next) if next == expected => Ok(MatchStatus::Match((&input[1..], next))),
-        _ => Ok(MatchStatus::NoMatch(input)),
-    }
-}
-
-pub fn numeric<'a>() -> impl Parser<'a, &'a str, char> {
-    move |input: &'a str| match input.chars().next() {
-        Some(next) if next.is_digit(10) => Ok(MatchStatus::Match((&input[1..], next))),
-        _ => Ok(MatchStatus::NoMatch(input)),
-    }
-}
